Ensure secret is generated or error Add --ensure-secret command line argument
