+++ .devcontainer/postCreateCommand.sh
@@ -18,3 +18,5 @@ poetry add mypy ruff black pytest --group=dev || true
 poetry install --with=dev || true
 
 mkdir -p .dev_container_logs
+
+poetry run ptyme-track --ensure-secret
