+++ ptyme_track/client.py
@@ -16,6 +16,7 @@ from ptyme_track.ptyme_env import (
     PTYME_TRACK_DIR,
     PTYME_WATCH_INTERVAL_MIN,
 )
+from ptyme_track.secret import validate_secret_file_exists
 from ptyme_track.server import sign_time
 from ptyme_track.signed_time import SignedTime
 
@@ -41,6 +42,7 @@ class PtymeClient:
         self._ignored_dirs = ignored_dirs
 
     def run_forever(self, cemented_file=Path(CEMENTED_PATH)) -> None:
+        validate_secret_file_exists()
         print("Starting ptyme-track", flush=True)
         prev_files_hash = None
         stopped = False
